+++ app-config.js
@@ -1,6 +1,6 @@
 globalThis.WC26_CONFIG = {
-  version: "704.14.13",
-  cacheName: "wc26-build-704-14-13",
+  version: "704.14.14",
+  cacheName: "wc26-build-704-14-14",
   versionUrl: "./version.json",
   updateCheckIntervalMs: 15 * 60 * 1000
 };